Allow injecting request context to the dry-view view
This adds a `view_context` setting to `WebPipe::Conn` which value is a
lambda taking the instance as argument. Unless an explicit `context:`
argument is given to the `#view` method, it is used as injected
arguments for `Dry::View::Context#with`.

For example:

```ruby
class MyContext < Dry::View::Context
  attr_reader :current_path

  def initialize(current_path: nil, **options)
    @current_path = current_path
    super
  end
end

class SayHelloView < Dry::View
  config.paths = [File.join(__dir__, '..', 'templates')]
  config.template = 'say_hello'
  config.default_context = MyContext.new

  expose :name
end

\#...
  def render(conn)
    conn.view(SayHelloView.new, name: 'Joe') # `current_path`
    # will be available in the view scope
  end
\#...
```