Allow overriding instance methods from the DSL
If we use `define_method` on the class that includes `WebPipe`, we lose
the possibility for it to call `super` (a new definition would
completely override ours). Instead, we now include a dynamic module.
That goes one level above the class in the ancestor's chain, so `super`
will work as expected.

For instance, someone might want to add some behavior to the initializer.

```ruby
class MyApp
  include WebPipe

  attr_reader :body

  def initialize(body:, **kwargs)
    @body = body
    super(**kwargs)
  end

  plug :render

  private

  def render(conn)
    conn.set_response_body(body)
  end
end
```

The same goes with any other instance method, like Rack's interface:

```ruby
class My App
  include WebPipe

  plug :render

  def render(conn)
    conn.set_response_body(conn.env['body'])
  end

  def call(env)
    env['body'] = 'Hello, world!'
    super
  end
end
```