Update hanami-view extension to work with the 2.1 beta
`Hanami::View::Context` no longer supports `#with`, as it's encouraged
for a single instance to be used for the whole request [1].

We use the same approach as hanami [2], where users can configure a
default context class and it's expected to be initialized with keyword
arguments. In our case, the default context class is a noop and we
expect users to implement it. It can be configured in the
`view_context_class` setting, while the initialize options can be
created from the connection struct through a lambda configured in the
`view_context_options` setting.

[1] - https://github.com/hanami/view/pull/223
[2] - https://github.com/hanami/hanami/pull/1359