Make web_pipe compatible with dry-auto_inject
We remove the built-in resolution of plugs from a container and instead
rely on the common mechanism provided by dry-auto_inject.

As we're already allowing the injection of plugs and middlewares through
keyword arguents given on `#initialize`, web_pipe remains compatible
with the keyword arguments strategy of dry-auto_inject. We make sure
that the order of inclusion of either web_pipe or dry-auto_inject is
irrelevant for a correct behavior via dispatching any extra keyword
arguments to `super`.

```ruby
WebPipe.load_extensions(:params)

class CreateUserApp
  include WebPipe
  include Deps[:create_user]

  plug :html, WebPipe::Plugs::ContentType.('text/html')
  plug :create

  private

  def create(conn)
    create_user.(conn.params)
  end
end
```